更新专辑，mq.媒体同步

+++ content/content_model/src/main/java/com/hls/content/po/TextInfo.java
@@ -29,27 +29,27 @@ public class TextInfo implements Serializable {
      * 文本ID（主键）
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 关联歌曲ID（歌词）
      */
-    private Integer songId;
+    private Long songId;
 
     /**
      * 关联歌手ID（歌手简介）
      */
-    private Integer singerId;
+    private Long singerId;
 
     /**
      * 关联专辑ID（专辑简介）
      */
-    private Integer albumId;
+    private Long albumId;
 
     /**
      * 创建用户ID
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 文本内容（歌词/简介）
